+++ app/controllers/admin/partners_controller.rb
@@ -1,4 +1,5 @@
 class Admin::PartnersController < ApplicationController
+  before_action :authenticate_user!
   before_action :authenticate_admin!
   before_action :set_partner, only: [:show, :edit, :update, :destroy]