+++ db/migrate/20151001172739_create_consent_forms.rb
@@ -1,0 +1,12 @@
+class CreateConsentForms < ActiveRecord::Migration
+  def change
+    create_table :consent_forms do |t|
+      t.string :title
+      t.string :name
+      t.text :content
+      t.boolean :deleted, default: false, null: false
+
+      t.timestamps null: false
+    end
+  end
+end