+++ test/models/admin/consent_form_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::ConsentFormTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end