+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Added admin dashboard
   - Ability to create team members for static page
   - Ability to create partners for static page
+  - Ability to create consent forms